requirements.txt
Django~=5.0
djangorestframework~=3.14.0
markdown~=3.5.1
django-filter~=23.5
Pillow
factory_boy
requests
pytz